--- android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/home/schooldetail/uistate/SchoolUiState.kt
@@ -1,6 +1,0 @@
-package com.festago.festago.presentation.ui.home.schooldetail.uistate
-
-data class SchoolUiState(
-    val id: Long,
-    val name: String,
-)
