+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/schooldetail/SchoolFestivalListAdapter.kt
@@ -1,9 +1,9 @@
-package com.festago.festago.presentation.ui.home.schooldetail
+package com.festago.festago.presentation.ui.schooldetail
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import com.festago.festago.presentation.ui.home.schooldetail.uistate.FestivalItemUiState
+import com.festago.festago.presentation.ui.schooldetail.uistate.FestivalItemUiState
 
 class SchoolFestivalListAdapter :
     ListAdapter<FestivalItemUiState, SchoolDetailFestivalViewHolder>(diffUtil) {
