+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/schooldetail/uistate/SchoolUiState.kt
@@ -1,0 +1,6 @@
+package com.festago.festago.presentation.ui.schooldetail.uistate
+
+data class SchoolUiState(
+    val id: Long,
+    val name: String,
+)
